Remove Strings class and its associated methods
